Summarize data in a table

Generates a Table 1 style summary of data, with an optional variable to stratify descriptions (overall summary statistics are always included). Data must be in long format.

Usage

summary(data, group_var, type)

Examples

# Load repeated measures of adolescent tooth growth (must be of long format)
data <- read.csv("https://raw.githubusercontent.com/alejandroh3005/modelLong/main/data/ortho.csv")[-1]
colnames(data)[3:4] <- c("Distance", "Age")

# Summarize data, stratify by sex, and ensure age is encoded as continuous
summary_res <- modelLong::summary(
  data = data,
  group_var = "Sex",
  type = list(Age ~ "continuous"))

# View summary table
summary_res$table1

  

Variable
      	Overall

N = 1621
      	
        
Sex

      
    

Female

N = 661
      	Male

N = 961
    

Subject

    F01
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F02
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F03
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F04
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F05
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F06
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F07
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F08
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F09
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F10
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    F11
	6 (3.7%)
	6 (9.1%)
	0 (0%)

    M01
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M02
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M03
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M04
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M05
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M06
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M07
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M08
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M09
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M10
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M11
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M12
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M13
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M14
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M15
	6 (3.7%)
	0 (0%)
	6 (6.3%)

    M16
	6 (3.7%)
	0 (0%)
	6 (6.3%)

Distance
	24.72 (22.50, 26.50)
	23.09 (21.50, 24.60)
	25.83 (23.50, 27.50)

Age
	13 (10, 16)
	13 (10, 16)
	13 (10, 16)

1 n (%); Mean (Q1, Q3)