Summarize data in a table

Generates a Table 1 style summary of data, with an optional variable to stratify descriptions (overall summary statistics are always included). Data must be in long format.

Usage

summary(data, group_var, type)

Examples

library(dplyr)

# Load repeated measures of adolescent tooth growth (must be of long format)
data <- read.csv(
  "https://raw.githubusercontent.com/alejandroh3005/modelLong/main/data/ortho.csv")[-1]
data <- data 

# Summarize data, stratify by sex, and ensure age is encoded as continuous
summary_res <- modelLong::summary(data,
                  group_var = "Sex",
                  type = list(Age ~ "continuous"))
#> Error in as_gt(.): Error processing `type` argument.
#> ! Can't select columns that don't exist. ✖ Column `Age` doesn't exist.
#> ℹ Select among columns "Subject", "distance", and "age"

# View summary table
summary_res$table1
#> Error: object 'summary_res' not found